[BUGFIX] Respect "0" max files setting when rotating logs
The documentation of `RotatingFileWriter` states [1] that setting
`maxFiles` to 0 keeps all log files after rotation, which isn't the
case right now.

This commit now checks whether the configuration is accordingly set,
including a new test to verify the bugfix.

[1] https://docs.typo3.org/permalink/t3coreapi:confval-rotating-file-writer-maxfiles

Resolves: #109636
Related: #100926
Releases: main, 13.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/93728
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>

+++ typo3/sysext/core/Classes/Log/Writer/RotatingFileWriter.php
@@ -112,9 +112,11 @@ class RotatingFileWriter extends FileWriter
         rsort($rotatedLogFiles, SORT_NATURAL);
 
         // Remove any excess files
-        $excessFiles = array_slice($rotatedLogFiles, $this->maxFiles);
-        foreach ($excessFiles as $excessFile) {
-            unlink($excessFile);
+        if ($this->maxFiles > 0) {
+            $excessFiles = array_slice($rotatedLogFiles, $this->maxFiles);
+            foreach ($excessFiles as $excessFile) {
+                unlink($excessFile);
+            }
         }
 
         $this->updateRuntimeRotationState(new \DateTimeImmutable());
